requirements.txt
anyascii==0.3.3
certifi==2025.10.5
charset-normalizer==3.4.4
defusedxml==0.7.1
et_xmlfile==2.0.0
filelock==3.20.0
fsspec==2025.9.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
langdetect==1.0.9
MarkupSafe==3.0.3
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
protobuf==6.33.0
pyclipper==1.3.0.post6
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-doctr==1.0.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
validators==0.35.0
